move more boilerplate to sourcegen
