+++ src/annot.rs
@@ -84,65 +84,12 @@ impl Atom {
   }
 }
 
-impl fmt::Display for Atom {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    let s = match self {
-      Atom::Str => "str",
-      Atom::Escape => "escape",
-      Atom::Hardbreak => "hardbreak",
-      Atom::Nbsp => "nbsp",
-      Atom::Blankline => "blankline",
-      Atom::ImageMarker => "image_marker",
-      Atom::LeftDoubleQuote => "left_double_quote",
-      Atom::RightDoubleQuote => "right_double_quote",
-      Atom::Ellipses => "ellipses",
-      Atom::Softbreak => "softbreak",
-      Atom::FootnoteReference => "footnote_reference",
-      Atom::OpenMarker => "open_marker",
-      Atom::Emoji => "emoji",
-      Atom::ReferenceKey => "reference_key",
-      Atom::ReferenceValue => "reference_value",
-      Atom::CodeLanguage => "code_language",
-      Atom::EmDash => "em_dash",
-      Atom::EnDash => "en_dash",
-    };
-    f.write_str(s)
-  }
-}
-
 impl Comp {
   pub(crate) fn add(self) -> Annot {
     Annot::Add(self)
   }
   pub(crate) fn sub(self) -> Annot {
     Annot::Sub(self)
-  }
-}
-
-impl fmt::Display for Comp {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    let s = match self {
-      Comp::Doc => "doc",
-      Comp::Verbatim => "verbatim",
-      Comp::Email => "email",
-      Comp::Url => "url",
-      Comp::Subscript => "subscript",
-      Comp::Superscript => "superscript",
-      Comp::Para => "para",
-      Comp::CodeBlock => "code_block",
-      Comp::Imagetext => "imagetext",
-      Comp::Linktext => "linktext",
-      Comp::Reference => "reference",
-      Comp::Destination => "destination",
-      Comp::Emph => "emph",
-      Comp::Strong => "strong",
-      Comp::DoubleQuoted => "double_quoted",
-      Comp::ReferenceDefinition => "reference_definition",
-      Comp::Insert => "insert",
-      Comp::Delete => "delete",
-      Comp::Mark => "mark",
-    };
-    f.write_str(s)
   }
 }
