+++ src/annot/generated.rs
@@ -1,3 +1,4 @@
+use std::fmt;
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub(crate) enum Comp {
   Doc,
@@ -21,9 +22,36 @@ pub(crate) enum Comp {
   Mark,
 }
 
+impl fmt::Display for Comp {
+  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+    f.write_str(match self {
+            Comp::Doc => "doc",
+      Comp::Verbatim => "verbatim",
+      Comp::Email => "email",
+      Comp::Url => "url",
+      Comp::Subscript => "subscript",
+      Comp::Superscript => "superscript",
+      Comp::Para => "para",
+      Comp::CodeBlock => "code_block",
+      Comp::Imagetext => "imagetext",
+      Comp::Linktext => "linktext",
+      Comp::Reference => "reference",
+      Comp::Destination => "destination",
+      Comp::Emph => "emph",
+      Comp::Strong => "strong",
+      Comp::DoubleQuoted => "double_quoted",
+      Comp::ReferenceDefinition => "reference_definition",
+      Comp::Insert => "insert",
+      Comp::Delete => "delete",
+      Comp::Mark => "mark",
+
+    })
+  }
+}
+    
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub(crate) enum Atom {
-  Str,
+    Str,
   Escape,
   Hardbreak,
   Nbsp,
@@ -41,6 +69,7 @@ pub(crate) enum Atom {
   CodeLanguage,
   EmDash,
   EnDash,
+
 }
 
 impl Atom {
@@ -51,3 +80,30 @@ impl Atom {
     matches!(self,  | Atom::RightDoubleQuote)
   }
 }
+
+impl fmt::Display for Atom {
+  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+    f.write_str(match self {
+            Atom::Str => "str",
+      Atom::Escape => "escape",
+      Atom::Hardbreak => "hardbreak",
+      Atom::Nbsp => "nbsp",
+      Atom::Blankline => "blankline",
+      Atom::ImageMarker => "image_marker",
+      Atom::LeftDoubleQuote => "left_double_quote",
+      Atom::RightDoubleQuote => "right_double_quote",
+      Atom::Ellipses => "ellipses",
+      Atom::Softbreak => "softbreak",
+      Atom::FootnoteReference => "footnote_reference",
+      Atom::OpenMarker => "open_marker",
+      Atom::Emoji => "emoji",
+      Atom::ReferenceKey => "reference_key",
+      Atom::ReferenceValue => "reference_value",
+      Atom::CodeLanguage => "code_language",
+      Atom::EmDash => "em_dash",
+      Atom::EnDash => "en_dash",
+
+    })
+  }
+}
+
